fix:修复端口号绑定失败的问题

+++ internal/k8s/frpc.go
@@ -71,7 +71,7 @@ func CreateFrpc(ctx context.Context, victim model.Victim) (model.Endpoints, []st
 			if slices.Contains(pr.Exclude, i) {
 				continue
 			}
-			portRange = append(portRange, i)
+			portRange = append(portRange, int32(i))
 		}
 	}
 	newEndpoints := make(model.Endpoints, 0)
